% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/myfn.R
\name{myfn}
\alias{myfn}
\title{Return arg + 1}
\usage{
myfn(tr_x, tr_y, te_x, te_y)
}
\arguments{
\item{tr_x}{train x}

\item{tr_y}{train y}

\item{te_x}{test x}

\item{te_y}{test y}
}
\description{
Return arg + 1
}

